requirements.txt
SpeechRecognition
pyaudio
gmusicapi
python-vlc
